Release 1.129.0

+++ classes/entities/quizquestionrelations/migration.php
@@ -49,7 +49,7 @@ class migration extends \local_intellidata\entities\migration {
      * @return array
      */
     public function get_sql($count = false, $condition = null, $conditionparams = []) {
-        $release4 = (float)ParamsHelper::get_release() >= 4.0;
+        $release4 = ParamsHelper::get_release() >= 4.0;
 
         if ($release4) {
             $select = ($count) ?
